requirements.txt
streamlit==1.32.0
paddleocr==2.7.3
paddlepaddle==2.5.2
opencv-python-headless
pillow
numpy
beautifulsoup4
lxml
